hub/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.7.0
pydantic-settings>=2.3.0
structlog>=24.1.0
alembic>=1.13.0
python-multipart>=0.0.9
websockets>=12.0
